requirements.txt
Django==3.1.7
djangorestframework==3.12.2
psycopg2==2.9.3
PyJWT==2.0.1
pyyaml==5.3.1